+++ CHANGES.rst
@@ -6,6 +6,7 @@ Here you can see the full list of changes between each hmmlearn release.
 next
 ----
 
+- Fixed support for multi-sequence GMM-HMM fit.
 - Deprecated ``utils.iter_from_X_lengths``.
 - Previously, APIs taking a *lengths* parameter would silently drop the last
   samples if the total length was less than the number of samples.  This